Generate for ASTNode subclasses constructors with all optional named parameters.
R=danrubel@google.com,brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org//12181015

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@18082 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/editor/tools/plugins/com.google.dart.engine_test/src/com/google/dart/engine/ast/ASTFactory.java
@@ -20,6 +20,7 @@ import com.google.dart.engine.utilities.dart.ParameterKind;
 
 import static com.google.dart.engine.scanner.TokenFactory.token;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -571,7 +572,7 @@ public final class ASTFactory {
   }
 
   public static IntegerLiteral integer(long value) {
-    return new IntegerLiteral(token(TokenType.INT, Long.toString(value)), value);
+    return new IntegerLiteral(token(TokenType.INT, Long.toString(value)), BigInteger.valueOf(value));
   }
 
   public static InterpolationExpression interpolationExpression(Expression expression) {
